cleanup mempool tests
